% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{Plotting multiple time series}
\usage{
plot_series(df, x, y, color, method = "ggplot", ...)
}
\arguments{
\item{df}{data frame with series data}

\item{x}{date variable (x axis)}

\item{y}{numeric variable (y axis)}

\item{color}{grouping variable}

\item{method}{either "ggplot" or "plotly"}

\item{...}{xlab, ylab, title or labels for time series}
}
\description{
Producing a static (ggplot) or dynamic (plotly) plot for visual comparison of several time series.
}
