% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Zscore}
\usage{
zscore(x, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{A logical value indicating whether missing values should be removed}
}
\description{
Compute z-scores
}
